Properly implemented the CLI
