+++ src/main/java/CLI/Commands/AddToRecipeBookCommand.java
@@ -1,4 +1,4 @@
-package main.java.CLI.Pages.Commands;
+package main.java.CLI.Commands;
 
 import main.java.CLI.CommandLineInterface;
 import main.java.Entities.Recipe;
