+++ src/main/java/CLI/Commands/Command.java
@@ -1,4 +1,4 @@
-package main.java.CLI.Pages.Commands;
+package main.java.CLI.Commands;
 
 import main.java.CLI.CommandLineInterface;
