+++ src/main/java/CLI/Commands/UpdatePreferencesCommand.java
@@ -1,0 +1,18 @@
+package main.java.CLI.Commands;
+
+import main.java.CLI.CommandLineInterface;
+
+/**
+ * Allows the user to update their preferences
+ */
+public class UpdatePreferencesCommand extends Command {
+    public UpdatePreferencesCommand() {
+        super("update preferences", "Starts a questionnaire to update user preferences");
+    }
+
+    @Override
+    public void runAction(CommandLineInterface CLI) {
+
+    }
+}
+
