+++ src/main/java/CLI/PageManager.java
@@ -1,0 +1,70 @@
+package main.java.CLI;
+
+import main.java.CLI.Commands.*;
+
+public class PageManager {
+    Page signedOut = new Page(
+            null,
+            new Command[]{
+                    new SignInCommand(),
+                    new CreateAccountCommand()
+            }
+    );
+
+    Page signedIn = new Page(
+            signedOut,
+            new Command[] {
+                    new EnterRecipeBookCommand(),
+                    new EnterRecipeViewerCommand(),
+                    new SignOutCommand(),
+                    new UpdatePreferencesCommand()
+            }
+    );
+
+    Page recipeViewer = new Page(
+            signedIn,
+            new Command[]  {
+                    new GetNewRecipeCommand(),
+                    new RateRecipeCommand(),
+                    new AddToRecipeBookCommand(),
+                    new GoBackCommand()
+            }
+    );
+
+    Page recipeBook = new Page(
+            signedIn,
+            new Command[] {
+                    new ListRecipeBookCommand(),
+                    new RemoveRecipeCommand(),
+                    new RateRecipeCommand(),
+                    new GoBackCommand()
+            }
+    );
+
+    Page currentPage = signedOut;
+
+    public void setPage(Page page) {
+        currentPage = page;
+    }
+    public Command findCommand(String name) {
+        return currentPage.findCommand(name);
+    }
+    public void signIn() {
+        setPage(signedIn);
+    }
+    public void signOut() {
+        setPage(signedOut);
+    }
+    public void enterRecipeBook() {
+        setPage(recipeBook);
+    }
+    public void enterRecipeViewer() {
+        setPage(recipeViewer);
+    }
+    public void goBack() {
+        setPage(currentPage.getParent());
+    }
+    public Command[] getAvailableCommands() {
+        return currentPage.getAvailableCommands();
+    }
+}
