--- src/main/java/CLI/Pages/Commands/QuitCommand.java
@@ -1,17 +1,0 @@
-package main.java.CLI.Pages.Commands;
-
-import main.java.CLI.CommandLineInterface;
-
-/**
- * Allows the user to quit the program
- */
-public class QuitCommand extends Command {
-    public QuitCommand() {
-        super("quit", "Quits the program");
-    }
-
-    @Override
-    public void runAction(CommandLineInterface CLI) {
-        CLI.isRunning = false;
-    }
-}
